core_services/vector_store/requirements.txt
qdrant-client==1.12.0
grpcio==1.66.1
grpcio-tools==1.66.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-grpc==0.48b0
prometheus_client==0.21.0
pytest==8.3.3
pytest-asyncio==0.24.0